\name{dbvsmsvm-package}
\alias{dbvsmsvm-package}
\docType{package}
\title{
A R Package to implement derivative-based variable selection on the multicategory support vector machines
}
\description{
Implementation of derivative-based variable selection on the multicategory support vector machines.
}

\details{
\tabular{rl}{
Package: \tab dbvsmsvm\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2021-02-22\cr
License: \tab GPL-2\cr
}
}

\author{
Beomjin Park

Maintainer : Beomjin Park <bbbeomjin@gmail.com>
}



%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
library(devtools)
install_github("bbeomjin/dbvsmsvm")
library(dbvsmsvm)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.